Created the Admin Dashboard

+++ routes/web.php
@@ -4,7 +4,6 @@ use App\Http\Controllers\ProductController;
 use App\Http\Controllers\CartController;
 use App\Http\Controllers\OrderController;
 use App\Http\Middleware\AdminMiddleware;
-use Illuminate\Foundation\Application;
 use Illuminate\Support\Facades\Route;
 use Inertia\Inertia;
 
@@ -12,11 +11,8 @@ Route::get('/', function () {
     return Inertia::render('Welcome', [
         'canLogin' => Route::has('login'),
         'canRegister' => Route::has('register'),
-        'laravelVersion' => Application::VERSION,
-        'phpVersion' => PHP_VERSION,
     ]);
 });
-
 
 Route::middleware([
     'auth:sanctum',
@@ -26,11 +22,11 @@ Route::middleware([
     
     Route::get('/products', [ProductController::class, 'index'])->name('products');
     Route::get('/products/{id}', [ProductController::class, 'show'])->name('products.show');
-    Route::post('/products', [ProductController::class, 'store'])->name('products.store');
-    Route::put('/products/{id}', [ProductController::class, 'update'])->name('products.update');
-    Route::delete('/products/{id}', [ProductController::class, 'destroy'])->name('products.destroy');
+    Route::post('/products', [ProductController::class, 'store'])->name('products.store')->middleware(AdminMiddleware::class);
+    Route::put('/products/{id}', [ProductController::class, 'update'])->name('products.update')->middleware(AdminMiddleware::class);
+    Route::delete('/products/{id}', [ProductController::class, 'destroy'])->name('products.destroy')->middleware(AdminMiddleware::class);
 
-    Route::get('/orders', [OrderController::class, 'index'])->name('orders');
+    Route::get('/orders', function () {return Inertia::render('Orders');})->name('orders');
     Route::get('/orders/{id}', [OrderController::class, 'show'])->name('orders.show');
     Route::post('/orders', [OrderController::class, 'store'])->name('orders.store');
     Route::put('/orders/{id}', [OrderController::class, 'update'])->name('orders.update');
@@ -40,9 +36,8 @@ Route::middleware([
     Route::post('/cart/add', [CartController::class, 'add'])->name('cart.add');
     Route::post('/cart/remove', [CartController::class, 'remove'])->name('cart.remove');
     Route::post('/cart/update', [CartController::class, 'update'])->name('cart.update');
+
+    Route::get('/dashboard', function () {
+        return Inertia::render('Dashboard');
+    })->name('dashboard')->middleware(AdminMiddleware::class);
 });
-
-
-// Route::middleware([AdminMiddleware::class])->group(function () {
-//     Route::post('/api/products', [ProductController::class, 'store']);
-// });
